refactor(style): improve formatting and variable names
