+++ src/components/Shell/RoomShareDialog.tsx
@@ -55,21 +55,26 @@ export function RoomShareDialog(props: RoomShareDialogProps) {
       props.roomId,
       password
     )
+
     if (encoded === props.password) {
       const params = new URLSearchParams()
       params.set('secret', props.password)
+
       await props.copyToClipboard(
         `${url}#${params}`,
         'Private room URL with password copied to clipboard',
         'warning'
       )
+
       handleClose()
     } else {
       setPassThrottled(true)
       props.showAlert('Incorrect password entered. Please wait 2s to retry.', {
         severity: 'error',
       })
+
       await sleep(2000)
+
       setPassThrottled(false)
     }
   }
@@ -82,11 +87,13 @@ export function RoomShareDialog(props: RoomShareDialogProps) {
         : 'Current URL copied to clipboard',
       'success'
     )
+
     handleClose()
   }
 
   const handleFormSubmit = (event: SyntheticEvent<HTMLFormElement>) => {
     event.preventDefault()
+
     if (!passThrottled) copyWithPass()
   }
