+++ src/index.tsx
@@ -2,9 +2,9 @@ import './polyfills'
 import ReactDOM from 'react-dom/client'
 import 'typeface-roboto'
 
-import 'index.sass'
-import Init from 'Init'
-import reportWebVitals from 'reportWebVitals'
+import './index.sass'
+import Init from './Init'
+import reportWebVitals from './reportWebVitals'
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(<Init />)
